Use on instance of react tooltip in Tutorial widget
