+++ src/components/tutorial/tutorial-tooltip/tutorial-tooltip.js
@@ -1,0 +1,8 @@
+import Component from './tutorial-tooltip-component';
+import { connect } from 'react-redux';
+import tutorialActions from 'redux_modules/tutorial';
+
+const actions = { ...tutorialActions };
+
+
+export default connect(null, actions)(Component);
