+++ src/constants/tutorial.js
@@ -2,3 +2,9 @@
 export const RARITY_RICHNESS_TUTORIAL = 'rarity-richness-tutorial';
 export const LEGEND_TUTORIAL = 'legend-tutorial';
 export const LEGEND_DRAG_TUTORIAL = 'legend-drag-tutorial';
+
+export const tutorialData = {
+  [RARITY_RICHNESS_TUTORIAL]: 'When activating a new biodiversity layer you can switch between rarity and richness.',
+  [LEGEND_TUTORIAL]: 'In the legend you can get the layer’s metadata info, change its opacity or remove it from the map.',
+  [LEGEND_DRAG_TUTORIAL]: 'Drag the legends to change its order on the map.',
+}
