+++ src/styles/common-animations.module.scss
@@ -25,13 +25,13 @@ $animation-time: 400ms;
   display: block;
   border-radius: 50%;
   cursor: pointer;
-  background-color: 0 0 0 rgba(255,255,255,0.4);
+  box-shadow: 0 0 0 rgba(255,255,255,0.4);
   animation: pulse 2s infinite;
-  transform: translateZ(0);
+  // transform: translateZ(0);
 }
 
 @keyframes pulse {
-  0% { background-color: 0 0 0 0 rgba(255,255,255,0.4); }
-  70% { background-color: 0 0 0 10px rgba(255,255,255,0); }
-  100% { background-color: 0 0 0 0 rgba(255,255,255, 0); }
+  0% { box-shadow: 0 0 0 0 rgba(255,255,255,0.4); }
+  70% { box-shadow: 0 0 0 10px rgba(255,255,255,0); }
+  100% { box-shadow: 0 0 0 0 rgba(255,255,255, 0); }
 }
